+++ CHANGELOG.md
@@ -32,6 +32,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - `client.batch.hisWrite(payload, options)`
   - `client.batch.hisRead(ids, from, to, options)`
   - `client.batch.create(entities, options)`
+  - `client.batch.hisDelete(ids, range, options)`
 
 ### CHANGED
 - Client no longer throws a Axios error if a response has been received and response is a Haystack of GraphQL error.